# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Travel planner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Trip {
  tripId: number;
  userId: string;
  title: string;
  tripDate: string;
  /** @nullable */
  memo?: string | null;
  /** @nullable */
  companions?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface TripInput {
  /** @maxLength 200 */
  title: string;
  tripDate: string;
  memo?: string;
  companions?: string;
}

export interface TripUpdate {
  /** @maxLength 200 */
  title?: string;
  tripDate?: string;
  /** @nullable */
  memo?: string | null;
  /** @nullable */
  companions?: string | null;
}

export interface CategoryBudgetExpense {
  categoryId: number;
  categoryName: string;
  budget: number;
  expense: number;
}

export interface TripSummary {
  tripId: number;
  title: string;
  tripDate: string;
  totalBudget: number;
  totalExpense: number;
  remaining: number;
  categoryBreakdown: CategoryBudgetExpense[];
}

export interface Dashboard {
  upcomingTrips: number;
  pastTrips: number;
  totalTrips: number;
  totalExpenseAllTime: number;
  recentTrips: Trip[];
}

export interface Transportation {
  transportationId: number;
  tripId: number;
  transportationType: string;
  departurePlace: string;
  arrivalPlace: string;
  departureTime: string;
  arrivalTime: string;
  fare: number;
}

export interface TransportationInput {
  /** @maxLength 50 */
  transportationType: string;
  /** @maxLength 100 */
  departurePlace: string;
  /** @maxLength 100 */
  arrivalPlace: string;
  departureTime: string;
  arrivalTime: string;
  fare: number;
}

export interface TransportationUpdate {
  /** @maxLength 50 */
  transportationType?: string;
  /** @maxLength 100 */
  departurePlace?: string;
  /** @maxLength 100 */
  arrivalPlace?: string;
  departureTime?: string;
  arrivalTime?: string;
  fare?: number;
}

export interface Accommodation {
  accommodationId: number;
  tripId: number;
  accommodationName: string;
  address: string;
  checkIn: string;
  checkOut: string;
  /** @nullable */
  reservationNumber?: string | null;
}

export interface AccommodationInput {
  /** @maxLength 200 */
  accommodationName: string;
  /** @maxLength 255 */
  address: string;
  checkIn: string;
  checkOut: string;
  reservationNumber?: string;
}

export interface AccommodationUpdate {
  /** @maxLength 200 */
  accommodationName?: string;
  /** @maxLength 255 */
  address?: string;
  checkIn?: string;
  checkOut?: string;
  /** @nullable */
  reservationNumber?: string | null;
}

export interface Budget {
  budgetId: number;
  tripId: number;
  categoryId: number;
  categoryName: string;
  budgetAmount: number;
}

export interface BudgetInput {
  categoryId: number;
  budgetAmount: number;
}

export interface BudgetUpdate {
  budgetAmount?: number;
}

export interface Expense {
  expenseId: number;
  tripId: number;
  categoryId: number;
  categoryName: string;
  expenseAmount: number;
  expenseDate: string;
  /** @nullable */
  paymentMethod?: string | null;
}

export interface ExpenseInput {
  categoryId: number;
  expenseAmount: number;
  expenseDate: string;
  paymentMethod?: string;
}

export interface ExpenseUpdate {
  categoryId?: number;
  expenseAmount?: number;
  expenseDate?: string;
  /** @nullable */
  paymentMethod?: string | null;
}

export interface Category {
  categoryId: number;
  categoryName: string;
}

export interface Notification {
  notificationId: number;
  tripId: number;
  /** @nullable */
  reminder?: string | null;
  /** @nullable */
  notificationDatetime?: string | null;
  /** @nullable */
  notificationType?: string | null;
}

export interface NotificationInput {
  reminder?: string;
  notificationDatetime?: string;
  notificationType?: string;
}

export interface NotificationUpdate {
  /** @nullable */
  reminder?: string | null;
  /** @nullable */
  notificationDatetime?: string | null;
  /** @nullable */
  notificationType?: string | null;
}