# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Travel planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips/{tripId}/accommodation:
    get:
      operationId: listAccommodation
      tags: [accommodation]
      summary: List accommodations for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Accommodation"
    post:
      operationId: createAccommodation
      tags: [accommodation]
      summary: Add accommodation to a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccommodationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accommodation"
  /trips/{tripId}/accommodation/{accommodationId}:
    get:
      operationId: getAccommodation
      tags: [accommodation]
      summary: Get accommodation detail
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: accommodationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accommodation"
    patch:
      operationId: updateAccommodation
      tags: [accommodation]
      summary: Update accommodation
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: accommodationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccommodationUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accommodation"
    delete:
      operationId: deleteAccommodation
      tags: [accommodation]
      summary: Delete accommodation
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: accommodationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content

  # Budget
  /trips/{tripId}/budget:
    get:
      operationId: listBudget
      tags: [budget]
      summary: List budgets for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
    post:
      operationId: createBudget
      tags: [budget]
      summary: Create budget for a category
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
  /trips/{tripId}/budget/{budgetId}:
    patch:
      operationId: updateBudget
      tags: [budget]
      summary: Update budget
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: budgetId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
    delete:
      operationId: deleteBudget
      tags: [budget]
      summary: Delete budget
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: budgetId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content

  # Expenses
  /trips/{tripId}/expenses:
    get:
      operationId: listExpenses
      tags: [expense]
      summary: List expenses for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expense]
      summary: Record an expense
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /trips/{tripId}/expenses/{expenseId}:
    patch:
      operationId: updateExpense
      tags: [expense]
      summary: Update expense
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expense]
      summary: Delete expense
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content

  # Categories
...
  /trips/{tripId}/notifications:
    get:
      operationId: listNotifications
      tags: [notification]
      summary: List notifications for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
    post:
      operationId: createNotification
      tags: [notification]
      summary: Create notification for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /trips/{tripId}/notifications/{notificationId}:
    patch:
      operationId: updateNotification
      tags: [notification]
      summary: Update notification
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
    delete:
      operationId: deleteNotification
      tags: [notification]
      summary: Delete notification
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
components:
  schemas:
...
    Accommodation:
      type: object
      required: [accommodationId, tripId, accommodationName, address, checkIn, checkOut]
      properties:
        accommodationId:
          type: integer
        tripId:
          type: integer
        accommodationName:
          type: string
        address:
          type: string
        checkIn:
          type: string
          format: date-time
        checkOut:
          type: string
          format: date-time
        reservationNumber:
          type: ["string", "null"]
    AccommodationInput:
      type: object
      required: [accommodationName, address, checkIn, checkOut]
      properties:
        accommodationName:
          type: string
          maxLength: 200
        address:
          type: string
          maxLength: 255
        checkIn:
          type: string
          format: date-time
        checkOut:
          type: string
          format: date-time
        reservationNumber:
          type: string
    AccommodationUpdate:
      type: object
      properties:
        accommodationName:
          type: string
          maxLength: 200
        address:
          type: string
          maxLength: 255
        checkIn:
          type: string
          format: date-time
        checkOut:
          type: string
          format: date-time
        reservationNumber:
          type: ["string", "null"]
    Budget:
      type: object
      required: [budgetId, tripId, categoryId, categoryName, budgetAmount]
      properties:
        budgetId:
          type: integer
        tripId:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: string
        budgetAmount:
          type: number
    BudgetInput:
      type: object
      required: [categoryId, budgetAmount]
      properties:
        categoryId:
          type: integer
        budgetAmount:
          type: number
    BudgetUpdate:
      type: object
      properties:
        budgetAmount:
          type: number
    Expense:
      type: object
      required: [expenseId, tripId, categoryId, categoryName, expenseAmount, expenseDate]
      properties:
        expenseId:
          type: integer
        tripId:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: string
        expenseAmount:
          type: number
        expenseDate:
          type: string
          format: date
        paymentMethod:
          type: ["string", "null"]
    ExpenseInput:
      type: object
      required: [categoryId, expenseAmount, expenseDate]
      properties:
        categoryId:
          type: integer
        expenseAmount:
          type: number
        expenseDate:
          type: string
          format: date
        paymentMethod:
          type: string
    ExpenseUpdate:
      type: object
      properties:
        categoryId:
          type: integer
        expenseAmount:
          type: number
        expenseDate:
          type: string
          format: date
        paymentMethod:
          type: ["string", "null"]
...
    Notification:
      type: object
      required: [notificationId, tripId]
      properties:
        notificationId:
          type: integer
        tripId:
          type: integer
        reminder:
          type: ["string", "null"]
        notificationDatetime:
          type: ["string", "null"]
          format: date-time
        notificationType:
          type: ["string", "null"]
    NotificationInput:
      type: object
      properties:
        reminder:
          type: string
        notificationDatetime:
          type: string
          format: date-time
        notificationType:
          type: string
    NotificationUpdate:
      type: object
      properties:
        reminder:
          type: ["string", "null"]
        notificationDatetime:
          type: ["string", "null"]
          format: date-time
        notificationType:
          type: ["string", "null"]