# lib/api-zod/src/generated/types/accommodationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Travel planner API
 * OpenAPI spec version: 0.1.0
 */

export interface AccommodationInput {
  /** @maxLength 200 */
  accommodationName: string;
  /** @maxLength 255 */
  address: string;
  checkIn: Date;
  checkOut: Date;
  reservationNumber?: string;
}