# lib/api-zod/src/generated/types/budgetInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Travel planner API
 * OpenAPI spec version: 0.1.0
 */

export interface BudgetInput {
  categoryId: number;
  budgetAmount: number;
}