# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Travel planner API
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  expenseId: number;
  tripId: number;
  categoryId: number;
  categoryName: string;
  expenseAmount: number;
  expenseDate: Date;
  /** @nullable */
  paymentMethod?: string | null;
}